refactor(config): consolidate config modules to load/normalize/write seams
config.ts now owns JSONC reading, legacy-key migration, and precedence;
config-normalize.ts absorbs the type-coercion helpers; config-writer.ts
absorbs the configure/init write intents; the editing types fold into
the editing registry. 12 config modules become 7 with no behavior
change to precedence, env overrides, migration, or validation.

Refs #413
