--- extensions/config/config-utils.ts
@@ -1,74 +1,0 @@
-export function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-export function merge<T>(base: T, patch: unknown): T {
-  if (!isRecord(base) || !isRecord(patch)) return (patch ?? base) as T;
-  const out: Record<string, unknown> = { ...base };
-  for (const [key, value] of Object.entries(patch)) {
-    out[key] = isRecord(value) && isRecord(out[key]) ? merge(out[key], value) : value;
-  }
-  return out as T;
-}
-
-export function envBool(env: NodeJS.ProcessEnv, name: string): boolean | undefined {
-  const value = env[name];
-  if (value === undefined) return undefined;
-  return ["1", "true", "yes", "on"].includes(value.toLowerCase());
-}
-
-export function bool(value: unknown, fallback: boolean): boolean {
-  return typeof value === "boolean" ? value : fallback;
-}
-
-export function positiveInt(value: unknown, fallback: number): number {
-  return typeof value === "number" && Number.isInteger(value) && value > 0 ? value : fallback;
-}
-
-export function stringValue(value: unknown, fallback: string): string {
-  return typeof value === "string" && value.length > 0 ? value : fallback;
-}
-
-export function optionalString(value: unknown, fallback?: string): string | undefined {
-  return typeof value === "string" && value.length > 0 ? value : fallback;
-}
-
-export function validEnvVarName(value: string): boolean {
-  return /^[A-Za-z_][A-Za-z0-9_]*$/.test(value);
-}
-
-export function enumValue<T extends string>(value: unknown, allowed: readonly T[], fallback: T): T {
-  return typeof value === "string" && (allowed as readonly string[]).includes(value)
-    ? (value as T)
-    : fallback;
-}
-
-export function stringArray(value: unknown, fallback: string[]): string[] {
-  return Array.isArray(value) && value.every((item) => typeof item === "string") ? value : fallback;
-}
-
-export function stringMatrix(value: unknown, fallback: string[][]): string[][] {
-  return Array.isArray(value) &&
-    value.every(
-      (scope) =>
-        Array.isArray(scope) && scope.length > 0 && scope.every((item) => typeof item === "string"),
-    )
-    ? value
-    : fallback;
-}
-
-export function enumArray<T extends string>(
-  value: unknown,
-  allowed: readonly T[],
-  fallback: T[],
-): T[] {
-  return Array.isArray(value) && value.every((item) => allowed.includes(item as T))
-    ? (value as T[])
-    : fallback;
-}
-
-export function optionalStringArray(value: unknown): string[] | undefined {
-  return Array.isArray(value) && value.every((item) => typeof item === "string")
-    ? value
-    : undefined;
-}
